final_template/requirements.txt
django
pillow
django-crispy-forms
crispy-bootstrap5
gunicorn
psycopg2